+++ ptc_viewer/test/support/inspection_test_adapter.ex
@@ -3,7 +3,7 @@ defmodule PtcViewer.TestInspectionAdapter do
 
   def pin_inspection(_path, _trace_source) do
     {:error,
-     {:unsupported_inspection_schema_version, %{artifact_version: 4, supported_version: 6}}}
+     {:unsupported_inspection_schema_version, %{artifact_version: 4, supported_version: 7}}}
   end
 
   def conversation(_source, _run_id), do: {:error, :not_called}